+++ ext/TensorKitMooncakeExt/utility.jl
@@ -19,6 +19,7 @@ Mooncake.tangent_type(::Type{<:HomSpace}) = Mooncake.NoTangent
 @zero_derivative DefaultCtx Tuple{typeof(TensorOperations.tensorstructure), AbstractTensorMap, Int, Bool}
 
 @zero_derivative DefaultCtx Tuple{typeof(TensorOperations.tensorcontract_structure), AbstractTensorMap, Index2Tuple, Bool, AbstractTensorMap, Index2Tuple, Bool, Index2Tuple}
+@zero_derivative DefaultCtx Tuple{typeof(TensorKit.planaralloc_contract), Any}
 
 @zero_derivative DefaultCtx Tuple{typeof(TensorKit.planar_trace), TensorKit.FusionTreePair, Index2Tuple, Index2Tuple}
 @zero_derivative DefaultCtx Tuple{typeof(TensorKit.has_shared_permute), AbstractTensorMap, Index2Tuple}